Move category registration text generation to util
